+++ examples_beta/openpose3d/include/openpose3d/pointGrey.hpp
@@ -7,10 +7,10 @@
 
 // Following OpenPose `tutorial_wrapper/` examples, we create our own class inherited from WorkerProducer.
 // This worker:
-    // 1. Set hardware trigger and the buffer to get the latest obtained frame.
-    // 2. Read images from FLIR cameras.
-    // 3. Turn them into std::vector<cv::Mat>.
-    // 4. Return the resulting images wrapped into a std::shared_ptr<std::vector<Datum3D>>.
+// 1. Set hardware trigger and the buffer to get the latest obtained frame.
+// 2. Read images from FLIR cameras.
+// 3. Turn them into std::vector<cv::Mat>.
+// 4. Return the resulting images wrapped into a std::shared_ptr<std::vector<Datum3D>>.
 // The HW trigger + reading FLIR camera code is highly based on the Spinnaker SDK examples `AcquisitionMultipleCamera` and specially `Trigger`
 // (located in `src/`). See them for more details about the cameras.
 // See `examples/tutorial_wrapper/` for more details about inhering the WorkerProducer class.