+++ include/openpose/producer/webcamReader.hpp
@@ -24,7 +24,8 @@ namespace op
          * @param webcamResolution const Point<int> parameter which specifies the desired camera resolution.
          * @param fps Double parameter which specifies the desired camera frame rate.
          */
-        explicit WebcamReader(const int webcamIndex = 0, const Point<int>& webcamResolution = Point<int>{}, const double fps = 30.);
+        explicit WebcamReader(const int webcamIndex = 0, const Point<int>& webcamResolution = Point<int>{},
+                              const double fps = 30., const bool throwExceptionIfNoOpened = true);
 
         ~WebcamReader();
 
@@ -37,6 +38,7 @@ namespace op
     private:
         double mFps;
         long long mFrameNameCounter;
+        bool mThreadOpened;
         cv::Mat mBuffer;
         std::mutex mBufferMutex;
         std::atomic<bool> mCloseThread;