+++ src/openpose/producer/videoCaptureReader.cpp
@@ -5,14 +5,14 @@
 
 namespace op
 {
-    VideoCaptureReader::VideoCaptureReader(const int index) :
+    VideoCaptureReader::VideoCaptureReader(const int index, const bool throwExceptionIfNoOpened) :
         Producer{ProducerType::Webcam},
         mVideoCapture{index}
     {
         try
         {
             // assert: make sure video capture was opened
-            if (!isOpened())
+            if (throwExceptionIfNoOpened && !isOpened())
                 error("VideoCapture (webcam) could not be opened.", __LINE__, __FUNCTION__, __FILE__);
         }
         catch (const std::exception& e)