+++ src/openpose/utilities/flagsToOpenPose.cpp
@@ -67,9 +67,9 @@ namespace op
             // Avoid duplicates (e.g. selecting at the time camera & video)
             if (!imageDirectory.empty() && !videoPath.empty())
                 error("Selected simultaneously image directory and video. Please, select only one.", __LINE__, __FUNCTION__, __FILE__);
-            else if (!imageDirectory.empty() && webcamIndex != 0)
+            else if (!imageDirectory.empty() && webcamIndex > 0)
                 error("Selected simultaneously image directory and webcam. Please, select only one.", __LINE__, __FUNCTION__, __FILE__);
-            else if (!videoPath.empty() && webcamIndex != 0)
+            else if (!videoPath.empty() && webcamIndex > 0)
                 error("Selected simultaneously video and webcam. Please, select only one.", __LINE__, __FUNCTION__, __FILE__);
 
             // Get desired ProducerType
@@ -103,7 +103,26 @@ namespace op
             {
                 // cameraFrameSize
                 const auto webcamFrameSize = op::flagsToPoint(webcamResolution, "1280x720");
-                return std::make_shared<WebcamReader>(webcamIndex, webcamFrameSize, webcamFps);
+                if (webcamIndex >= 0)
+                {
+                    const auto throwExceptionIfNoOpened = true;
+                    return std::make_shared<WebcamReader>(webcamIndex, webcamFrameSize, webcamFps, throwExceptionIfNoOpened);
+                }
+                else
+                {
+                    const auto throwExceptionIfNoOpened = false;
+                    std::shared_ptr<WebcamReader> webcamReader;
+                    for (auto index = 0 ; index < 10 ; index++)
+                    {
+                        webcamReader = std::make_shared<WebcamReader>(index, webcamFrameSize, webcamFps, throwExceptionIfNoOpened);
+                        if (webcamReader->isOpened())
+                        {
+                            log("Auto-detecting camera index... Detected and opened camera " + std::to_string(index) + ".", Priority::High);
+                            return webcamReader;
+                        }
+                    }
+                    error("No camera found.", __LINE__, __FUNCTION__, __FILE__);
+                }
             }
             // else
             error("Undefined Producer selected.", __LINE__, __FUNCTION__, __FILE__);