+++ doc/release_notes.md
@@ -5,7 +5,7 @@ OpenPose Library - Release Notes
 
 ## OpenPose 1.0.0rc1
 1. Initial version, main functionality:
-    1. Body keypoint detection and rendering on Ubuntu 14 and 16.
+    1. Body keypoint detection and rendering in Ubuntu 14 and 16.
     2. It can read an image directory, video or webcam.
     3. It can display the results or storing them on disk.
 
@@ -90,6 +90,8 @@ OpenPose Library - Release Notes
     3. Added OpenPose 1.0.1 portable demo.
     4. Removed Python and some unnecessary boost dependencies on the VS project.
     5. Replaced all double quotes by angle brackets in include statements (issue #61).
+    6. Added 3-D reconstruction demo.
+    7. Auto-detection of the camera index.
 2. Main bugs fixed:
     1. Pycaffe can now be imported from Python.
     2. Fixed `Tutorial/Wrapper` VS linking errors.